Netbox Trust Boundary/requirements.txt
PySide6>=6.5.0
requests>=2.31.0
PyYAML>=6.0
